requirements.txt
# Core Dependencies
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.3.2

# Markov Chain & Probabilistic Models
markovify==0.9.4
transitions==0.9.0
pymc==5.10.0

# Data Validation & Quality
pandera==0.18.1
pydantic==2.5.2
great-expectations==0.18.7

# MLOps & Experiment Tracking
mlflow==2.10.2
dvc==3.46.1
prefect==2.14.10

# API & Serving
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0

# Monitoring & Observability
evidently==0.4.14
prometheus-client==0.19.0

# Visualization & Dashboard
streamlit==1.31.1
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Statistical Tests
statsmodels==0.14.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.67.0

# Testing & Code Quality
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Docker & Containerization
docker==7.0.0

# Development
jupyter==1.0.0
ipython==8.18.1

plotly-express==0.4.2
streamlit
pyyaml
matplotlib
seaborn

